refactor: 細かい修正
